Make PublicBundle store keys contiguously
This makes signing much easier, since we have a continuous chunk of data
we can easily sign and verify.

The disadvantage is that we now have to add functions to access the nth
public key in a bundle, and to tell us how many keys are in a bundle.

This also causes a divergence in API with PrivateBundle, which doesn't
need to use a contiguous chunk of data, and instead stores a slice of
private keys.